requirements_training.txt
# MS-Swift Training Requirements
# For training QwenVL-2B on custom datasets

# Core MS-Swift framework
ms-swift[all]>=3.0.0

# PyTorch (CUDA 11.8 or 12.1 recommended)
torch>=2.0.0
torchvision>=0.15.0

# Transformers and related
transformers>=4.40.0
accelerate>=0.28.0
peft>=0.10.0

# For Qwen2-VL specifically
qwen-vl-utils>=0.0.2

# DeepSpeed for multi-GPU training (optional)
deepspeed>=0.14.0

# Data processing
datasets>=2.18.0
pandas>=2.0.0
pyarrow>=14.0.0

# Image processing
Pillow>=10.0.0

# Progress and logging
tqdm>=4.66.0
tensorboard>=2.16.0
wandb>=0.16.0

# Utilities
pyyaml>=6.0
einops>=0.7.0
